Disable spina frontend
