--- app/decorators/controllers/spina/application_controller_decorator.rb
@@ -1,9 +1,0 @@
-Spina::ApplicationController.class_eval do
-  http_basic_authenticate_with name: ENV['HTTP_AUTH_USERNAME'], password: ENV['HTTP_AUTH_PASSWORD'] if Rails.env.production?
-
-  before_action :initialize_client
-
-  def initialize_client
-    @registers_client ||= RegistersClient::RegisterClientManager.new(cache_duration: 600)
-  end
-end
