+++ app/views/new_register/index.html.haml
