+++ app/views/new_register/suggest_register.html.haml
