+++ app/views/new_register/thanks.html.haml
