+++ app/views/pages/cookies.html.haml
@@ -1,0 +1,1 @@
+%h1 Cookies
