+++ app/views/registers/_record.html.haml
