+++ app/views/registers/_timeline_record.html.haml
