+++ app/views/registers/history.html.haml
