+++ app/views/registers/history.js.erb
