+++ app/views/registers/info.html.haml
@@ -1,0 +1,88 @@
+- @page_title = "#{@register.name.capitalize} Register"
+- @page_description = ""
+
+%nav{class: "breadcrumbs", "aria-label" => "Breadcrumbs"}
+  %ol{itemscope: "", itemtype: "http://schema.org/BreadcrumbList"}
+    %li{class: "breadcrumbs__item", itemprop: "itemListElement", itemscope: "", itemtype: "http://schema.org/ListItem"}
+      = link_to "https://www.gov.uk/service-toolkit#components", itemprop: "item" do
+        %span{itemprop: "name"} Components
+    %li{class: "breadcrumbs__item", itemprop: "itemListElement", itemscope: "", itemtype: "http://schema.org/ListItem"}
+      = link_to root_path, itemprop: "item" do
+        %span{itemprop: "name"} GOV.UK Registers
+    %li{class: "breadcrumbs__item", itemprop: "itemListElement", itemscope: "", itemtype: "http://schema.org/ListItem"}
+      = link_to registers_path, itemprop: "item" do
+        %span{itemprop: "name"} Register pipeline
+    %li{class: "breadcrumbs__item breadcrumbs__item", itemprop: "itemListElement", itemscope: "", itemtype: "http://schema.org/ListItem"}
+      = link_to register_path(@register.slug) do
+        %span{itemprop: "name"} #{@register.name} Register
+    %li{class: "breadcrumbs__item breadcrumbs__item--active", itemprop: "itemListElement", itemscope: "", itemtype: "http://schema.org/ListItem"}
+      %span{itemprop: "name"} Information
+
+%main.container{role:'main'}
+  .grid-row
+    .column-two-thirds
+      %h1= @page_title
+
+      - if @register.url.present?
+        %p #{link_to "View the register here", @register.url, target: :blank}. If you have a question or want to give feedback on the register #{link_to 'contact the team', support_path}.
+      .govspeak
+        = @register.history
+
+      - if @register.phases.any?
+        %h2.heading-large Register history
+        .phase-history
+          - @register.phases.each do |phase|
+            .phase-history-item
+              %h3.tag= phase.name
+              %p= phase.phase_update
+
+        %h2.heading-large Register updates
+        - @register.phases.each do |phase|
+          - grouped_steps = @register.steps.select { |step| step.step_phase == phase.name }
+
+          - grouped_steps.each do |step|
+            %h3.heading-medium{id: "#{step.title.parameterize}"}
+              %span.tag
+                = phase.name
+              = step.title
+            .govspeak
+              = step.content
+
+    .column-third
+      %aside#related
+        .related-container
+          .group
+            %h4.heading-small Managed by
+            %div{class: "govuk-organisation-logo #{@register.authority.parameterize}"}
+              %div{class: "logo-container #{crest_class_name(@register.authority.parameterize)}"}
+                %span= @register.authority
+            - if @register.custodian.present?
+              %h4.heading-small Custodian
+              %p= @register.custodian
+            - if @register.url.present?
+              %h4.heading-small Register
+              %p= link_to @register.url, @register.url, rel: 'external'
+
+            %hr
+
+            - @register.phases.each do |phase|
+              - grouped_steps = @register.steps.select { |step| step.step_phase == phase.name }
+
+              %h4.heading-small
+                %span.tag
+                  = phase.name
+
+              %ul.tick-list
+                - grouped_steps.each do |step|
+                  %li
+                    - if step.completed.present?
+                      %svg#Layer_1{"enable-background" => "new 0 0 35 35", :height => "30px", :space => "preserve", :version => "1.1", :viewbox => "0 0 35 35", :width => "30px", :x => "0px", :xmlns => "http://www.w3.org/2000/svg", "xmlns:xlink" => "http://www.w3.org/1999/xlink", :y => "0px"}
+                        %g
+                          %g
+                            %path{:d => "M12.393,29.908", :fill => "#231F20"}
+                        %g
+                          %polyline{:fill => "#FFFFFF", :points => "30.848,7.857 12.811,25.895 3.836,16.921 \t"}
+                          %polygon{:fill => "#231F20", :points => "27.012,11.66 10.742,27.93 3.535,20.724 0,24.26 10.742,35 30.547,15.195 \t"}
+                    - else
+                      %svg#Layer_1{"enable-background" => "new 0 0 35 35", :height => "30px", :space => "preserve", :version => "1.1", :viewbox => "0 0 35 35", :width => "30px", :x => "0px", :xmlns => "http://www.w3.org/2000/svg", "xmlns:xlink" => "http://www.w3.org/1999/xlink", :y => "0px"}
+                    = link_to step.title, "#{ '#' + step.title.parameterize}"
