+++ app/views/registers/show.html.haml
