+++ app/views/registers/show.js.erb
