--- app/views/spina/admin/page_partables/texts/_form.html.haml
@@ -1,5 +1,0 @@
-.horizontal-form-label
-  = f.object.title
-.horizontal-form-content
-  = f.fields_for :page_partable, f.object.page_partable do |ff|
-    = ff.text_area :content
