+++ app/views/support/thanks.html.haml
